+++ scripts/website_screenshots.js
@@ -5,17 +5,19 @@ const puppeteer = require('puppeteer');
   const page = await browser.newPage()
   await page.setViewport({ width: 1920, height: 1080 })
   await page.goto('http://localhost:8080')
-
   await page.click('#theme_select')
   await connectWithServer(page)
   await removeSnackbar(page)
-  await clickToNavigateAndScreenshot(page, '#navbar_home', 'screenshot_0_home_white.jpg')
+  await page.click('i.mdi-close-circle')
+
+  await clickToNavigateAndScreenshot(page, '#navbar_home', 'screenshot_0_home_white.png')
   await page.click('#theme_select')
-  await clickToNavigateAndScreenshot(page, '#navbar_nodes', 'screenshot_1_nodes.jpg')
-  await clickToNavigateAndScreenshot(page, ['#navbar_indices', 'button[title="Options"]'], 'screenshot_2_indices.jpg')
+  await clickToNavigateAndScreenshot(page, '#navbar_nodes', 'screenshot_1_nodes.png')
+  await clickToNavigateAndScreenshot(page, ['#navbar_indices', 'button[title="Options"]'], 'screenshot_2_indices.png')
+  await clickToNavigateAndScreenshot(page, ['#navbar_indices', 'button[title="Options"]', 'i.mdi-at'], 'screenshot_3_indices.png')
   await page.reload()
   await page.click('#theme_select')
-  await clickToNavigateAndScreenshot(page, '#navbar_search', 'screenshot_4_search_dark.jpg', async page => {
+  await clickToNavigateAndScreenshot(page, '#navbar_search', 'screenshot_4_search_dark.png', async page => {
     await page.click('#index-pattern')
     await page.waitForTimeout(50)
     await page.click('input[type="checkbox"]')
@@ -24,14 +26,15 @@ const puppeteer = require('puppeteer');
     await page.click('th[aria-label="author_name (author_name.keyword): Not sorted. Activate to sort ascending."')
     await page.waitForTimeout(250)
   })
-  await clickToNavigateAndScreenshot(page, '#navbar_query_rest', 'screenshot_5_query_dark.jpg', async page => {
+  await clickToNavigateAndScreenshot(page, '#navbar_query_rest', 'screenshot_5_query_dark.png', async page => {
     await page.focus('#path')
     await page.keyboard.type('_search')
     await page.click('#execute_query')
     await page.waitForTimeout(500)
   })
-  await clickToNavigateAndScreenshot(page, ['#navbar_snapshots', '#navbar_snapshots_repositories', 'table tbody tr.tr--clickable'], 'screenshot_6_snapshots_dark.jpg')
-  await clickToNavigateAndScreenshot(page, '#navbar_utilities', 'screenshot_7_utilities_dark.jpg')
+  await clickToNavigateAndScreenshot(page, ['#navbar_snapshots', '#navbar_snapshots_repositories', 'table tbody tr.tr--clickable'], 'screenshot_6_snapshots_dark.png')
+  await clickToNavigateAndScreenshot(page, '#navbar_utilities', 'screenshot_7_utilities_dark.png')
+  await clickToNavigateAndScreenshot(page, '#navbar_settings', 'screenshot_8_settings.png')
 
   await browser.close()
 })()
@@ -61,7 +64,7 @@ async function clickToNavigateAndScreenshot (page, selectors, screenshot, callba
         await page.click(selector)
       } catch (e) {
         console.log('error.', e)
-        await page.screenshot({ path: 'scripts/website_screenshots/error.jpg' })
+        await page.screenshot({ path: 'scripts/website_screenshots/error.png' })
       }
     }
   } else {