+++ src/components/Indices/IndexAliases.vue
@@ -45,7 +45,7 @@
         </v-form>
       </v-card-text>
 
-      <v-data-table :footer-props="{itemsPerPageOptions: DEFAULT_ITEMS_PER_PAGE}"
+      <v-data-table :footer-props="{itemsPerPageOptions: DEFAULT_ITEMS_PER_PAGE, showFirstLastPage: true}"
                     :headers="HEADERS"
                     :items="items || []"
                     :loading="requestState.loading"