+++ src/components/shared/ChangeTheme.vue
@@ -5,12 +5,12 @@
 
 <script>
   import { watch } from '@vue/composition-api'
-  import { compositionVuexAccessors } from '@/helpers/store'
+  import { vuexAccessors } from '@/helpers/store'
 
   export default {
     name: 'change-theme',
     setup (props, context) {
-      const { dark } = compositionVuexAccessors('theme', ['dark'])
+      const { dark } = vuexAccessors('theme', ['dark'])
 
       watch(dark, newValue => {
         context.root.$vuetify.theme.dark = newValue