+++ src/components/shared/CodeViewer.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script>
-  import { compositionVuexAccessors } from '@/helpers/store'
+  import { vuexAccessors } from '@/helpers/store'
   import { onBeforeUnmount, onMounted, ref, watch } from '@vue/composition-api'
   import store from '@/store'
   import BtnGroup from '@/components/shared/BtnGroup'
@@ -41,7 +41,7 @@
     },
     setup (props, context) {
       const editor = ref(null)
-      const { useSpaces, wrapLines } = compositionVuexAccessors('codeEditor', ['useSpaces', 'wrapLines'])
+      const { useSpaces, wrapLines } = vuexAccessors('codeEditor', ['useSpaces', 'wrapLines'])
       const { setTheme, setWhitespace, setWrapLines, unmountEditor, setupAceEditor } = editorUtils(editor)
 
       watch(() => props.value, newValue => {